Fix illegal bid when opener asks Blackwood directly over Jacoby 2NT
After 1M-2NT (Jacoby), an opener rebid of 4NT fell into the generic
"Blackwood or return to game" branch, whose "Returning to game" rule bid
4M below the current 4NT contract — an illegal call. Route opener's 4NT
to blackwoodAnswerRules() instead, so the responder answers aces and the
existing blackwoodPlacementRules complete the slam auction.

Also add a general safety net: selectSaycBid now skips any rule whose
action is not legal in the current auction, and legality-checks the
fallback bidder's choice, so no unanticipated human call can ever make
the engine choose an illegal bid again. isLegalCall moves from
selfplay.dart into sayc_bidding.dart to support this (script imports
updated).

Verified: 301 bridge tests pass; 3000-deal self-play audit shows no
hard failures and unchanged heuristic-flag counts.
