+++ scripts/bridge_play_compare.dart
@@ -28,7 +28,7 @@ import 'package:cards_with_cats/bridge/bridge.dart';
 import 'package:cards_with_cats/bridge/bridge_ai.dart';
 import 'package:cards_with_cats/bridge/play_strategies.dart';
 import 'package:cards_with_cats/bridge/sayc/sayc_bidding.dart';
-import 'package:cards_with_cats/bridge/sayc/selfplay.dart' show isLegalCall;
+import 'package:cards_with_cats/bridge/sayc/sayc_bidding.dart' show isLegalCall;
 
 class StrategyStats {
   int cardsPlayed = 0;
